pyproject.toml
[project]
name = "plc-tester"
version = "1.0.0"
description = "Professional PLC Siemens S7-1200/1500 diagnostic and communication testing tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6",
    "python-snap7>=1.3",
    "asyncua>=1.0",
]

[project.scripts]
plc-tester = "plc_tester.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/plc_tester"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "ruff>=0.4",
    "pytest>=8.0",
]
